refactor(api): move organization-invitation serializer
